Assignment 5.1 - Updates to Final
Updated the Group Boxes for Pre-Tax and Post-Tax Deductions to Lists, Updated the Gross Income Textbox to accept only numerical values and only one decimal point. I used the string ToUpper to capitalize the name in the welcome banner. I used String Replace to replace the TBs from the Textbox names so that the list items would be recognized for the adding function and real-time update to the totals.